spostate robe in stdlib, sistemati conflitti vari, tweaks e commenti vari
